requirements.txt
streamlit
opencv-python>=4.10.0
tensorflow-cpu
